+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 - General
+  - Consider `.lsp/config.edn` as part of project code, removing false positives of unused-public-var linter.
+  - Consider full qualified symbols in edn files when checking for var references.
   - New feature: Add support for custom project linters. #2043
 
 - Editor